+++ src/main/java/org/folio/inventory/domain/ingest/IngestJobCollection.java
@@ -1,9 +1,9 @@
 package org.folio.inventory.domain.ingest;
 
 import org.folio.inventory.domain.AsynchronousCollection;
-import org.folio.inventory.resources.ingest.IngestJob;
+import org.folio.rest.jaxrs.model.IngestStatus;
 
 public interface IngestJobCollection
-  extends AsynchronousCollection<IngestJob> {
+  extends AsynchronousCollection<IngestStatus> {
 
 }